Draft: wifi: Assorted PHY, MAC and rate-control fixes (fixes #424, #649, #746, #1022, #1090, #1166, #1194, #1199; refs #414)

Closes #424, #649, #746, #1022, #1090, #1166, #1194, and #1199

See merge request nsnam/ns-3-dev!2906

+++ src/wifi/model/rate-control/thompson-sampling-wifi-manager.h
@@ -109,7 +109,7 @@ class ThompsonSamplingWifiManager : public WifiRemoteStationManager
      * @param beta second parameter of beta distribution
      * @return beta random variable sample
      */
-    double SampleBetaVariable(uint64_t alpha, uint64_t beta) const;
+    double SampleBetaVariable(double alpha, double beta) const;
 
     Ptr<GammaRandomVariable>
         m_gammaRandomVariable; //!< Variable used to sample beta-distributed random variables
